# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Banque",
        "description": "API pour la gestion des comptes bancaires",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur API"
        }
    ],
    "paths": {
        "/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister tous les comptes",
                "description": "Admin peut récupérer la liste de tous les comptes, Client peut récupérer la liste de ses comptes. Liste des comptes non supprimés, type cheque ou epargne, actifs.",
                "operationId": "5fdb1d9d6a1e1d81e65a694d33a249c6",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "epargne",
                                "cheque"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "bloque",
                                "ferme"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par titulaire ou numéro",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/Pagination"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/Links"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/comptes/{compteId}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Afficher un compte spécifique",
                "description": "Admin peut récupérer un compte par ID, Client peut récupérer un de ses comptes par ID. Recherche locale par défaut, serverless si non trouvé.",
                "operationId": "53eb952653f2afcba77ec73e86d8ee69",
                "parameters": [
                    {
                        "name": "compteId",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "COMPTE_NOT_FOUND"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Le compte avec l'ID spécifié n'existe pas"
                                                },
                                                "details": {
                                                    "properties": {
                                                        "compteId": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Modifier un compte",
                "description": "Modifier les informations d'un compte bancaire existant",
                "operationId": "03d15875d7640e630f1a4d1f3352945e",
                "parameters": [
                    {
                        "name": "compteId",
                        "in": "path",
                        "description": "ID du compte à modifier",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type_compte": {
                                        "type": "string",
                                        "enum": [
                                            "epargne",
                                            "cheque"
                                        ],
                                        "example": "epargne"
                                    },
                                    "solde": {
                                        "type": "number",
                                        "example": 150000
                                    },
                                    "devise": {
                                        "type": "string",
                                        "example": "FCFA"
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": [
                                            "actif",
                                            "bloque",
                                            "ferme"
                                        ],
                                        "example": "actif"
                                    },
                                    "motifBlocage": {
                                        "type": "string",
                                        "example": "Raison du blocage"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte modifié avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Supprimer un compte",
                "description": "Supprimer un compte bancaire (soft delete)",
                "operationId": "36fe94063a393703c268b72fd33d18ce",
                "parameters": [
                    {
                        "name": "compteId",
                        "in": "path",
                        "description": "ID du compte à supprimer",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compte supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Compte": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "numeroCompte": {
                        "type": "string",
                        "example": "C00123456"
                    },
                    "titulaire": {
                        "type": "string",
                        "example": "Amadou Diallo"
                    },
                    "type": {
                        "type": "string",
                        "example": "epargne"
                    },
                    "solde": {
                        "type": "number",
                        "example": 1250000
                    },
                    "devise": {
                        "type": "string",
                        "example": "FCFA"
                    },
                    "dateCreation": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-03-15T00:00:00Z"
                    },
                    "statut": {
                        "type": "string",
                        "example": "bloque"
                    },
                    "motifBlocage": {
                        "type": "string",
                        "example": "Inactivité de 30+ jours"
                    },
                    "metadata": {
                        "properties": {
                            "derniereModification": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "version": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "properties": {
                    "currentPage": {
                        "type": "integer",
                        "example": 1
                    },
                    "totalPages": {
                        "type": "integer",
                        "example": 3
                    },
                    "totalItems": {
                        "type": "integer",
                        "example": 25
                    },
                    "itemsPerPage": {
                        "type": "integer",
                        "example": 10
                    },
                    "hasNext": {
                        "type": "boolean",
                        "example": true
                    },
                    "hasPrevious": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "Links": {
                "properties": {
                    "self": {
                        "type": "string",
                        "example": "/api/v1/comptes?page=1&limit=10"
                    },
                    "next": {
                        "type": "string",
                        "example": "/api/v1/comptes?page=2&limit=10"
                    },
                    "first": {
                        "type": "string",
                        "example": "/api/v1/comptes?page=1&limit=10"
                    },
                    "last": {
                        "type": "string",
                        "example": "/api/v1/comptes?page=3&limit=10"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Comptes",
            "description": "Comptes"
        }
    ],
    "security": [
        {
            "passport": []
        }
    ]
}